Add MMTYPE bitmasks
There are dozens of MMTYPEs in the specification, whose values are
spaced at intervals of four. This is becase a bitmask is used to
indicate whether the MME is a Request, Confirm, Indication or Response.
It makes little sense to list all four variants of each MMTYPE when they
are trivially calculated with the bitmask.