Further separate Qualcomm-specific code from the main code (#18)
* Refactor write_homeplug function as writeQualcommReq

Since the existing write_homeplug function only generates a Qualcomm
vendor-specific request, it makes sense to reflect that in the function
name. Make the MM type a parameter, so that the function is not limited
to NW_INFO.REQ packets.

* Align HomeplugNetworkStatus struct definition with reference

Align HomeplugNetworkStatus struct definition with open-plc-utils
reference implementation (cf. network struct, plc/NetInfo1.c). This
makes the decoding of a NW_INFO response a bit prettier and allows for a
simple binary.Read() to unmarshal the HomeplugNetworkStatus struct.

* Add MMTYPE bitmasks

There are dozens of MMTYPEs in the specification, whose values are
spaced at intervals of four. This is becase a bitmask is used to
indicate whether the MME is a Request, Confirm, Indication or Response.
It makes little sense to list all four variants of each MMTYPE when they
are trivially calculated with the bitmask.

* Drop net.Interface from read_homeplug function params

The interface MTU is irrelevant, since the HPAV specification limits
management messages to 1518 bytes. If the interface MTU is lower,
obviously fewer bytes will be read from the socket.

* Abstract raw socket read and Ethernet frame unmarshaling

Abstract code which will be common to handling all future MMTYPE
responses.

* Align struct member names / terminology  with HPAV specification

Essentially, MMEType -> MMType.

"MME" (Management Message Entry) describes the whole message, including
variable length fields, not merely the 16-bit type.